requirements.txt
flask>=3.0,<4
flask-cors>=4.0
python-dotenv>=1.0
huggingface_hub>=0.23
pdfplumber>=0.11
pdfminer.six>=20221105
numpy>=1.24,<2.0
scikit-learn>=1.3
pandas>=2.0
requests>=2.31
openpyxl>=3.1
pillow>=10.0
psutil>=5.9.0
